# 5.UNG_DUNG/5.1.backend_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HitRadar Pro API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/model-info": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Model Info",
        "description": "Return model metadata and test-set metrics.\n\nRaises 503 if the model is not loaded.",
        "operationId": "model_info_model_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfoResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/features": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Features",
        "description": "Return the 18 canonical input fields and 31 selected feature names.\n\nRaises 503 if the model is not loaded.",
        "operationId": "features_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeaturesResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/predict": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict",
        "description": "Predict track popularity from 18 audio features.\n\nReturns raw and clipped (0–100) predictions plus model metadata.\nRaises 503 if the model is not loaded.",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/explain": {
      "post": {
        "tags": [
          "explanation"
        ],
        "summary": "Explain",
        "description": "Predict and return SHAP feature attribution.\n\nComputes shap.TreeExplainer values for the 31 selected features.\nReturns top-5 features by absolute SHAP magnitude.\n\nIMPORTANT: SHAP values show feature importance, NOT causal relationships.\nRaises 503 if the model is not loaded.",
        "operationId": "explain_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/what-if": {
      "post": {
        "tags": [
          "what-if"
        ],
        "summary": "What If",
        "description": "Compare predictions before and after feature changes.\n\nSends base_features + changed_features dict.\nReturns before/after predictions and delta.\n\nRaises 503 if the model is not loaded.\nRaises 422 if any key in changed_features is not a canonical field name.",
        "operationId": "what_if_what_if_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "example": {
          "details": [],
          "error_code": "INVALID_FEATURE",
          "message": "Feature is not a valid canonical input field.",
          "timestamp": "2026-08-05T10:00:00Z"
        }
      },
...
      "FeaturesResponse": {
        "properties": {
          "canonical_fields": {
            "items": {
              "$ref": "#/components/schemas/FieldDescriptor"
            },
            "type": "array",
            "title": "Canonical Fields"
          },
          "selected_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Features"
          },
          "total_input_fields": {
            "type": "integer",
            "title": "Total Input Fields"
          },
          "total_selected_features": {
            "type": "integer",
            "title": "Total Selected Features"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "canonical_fields",
          "selected_features",
          "total_input_fields",
          "total_selected_features"
        ],
        "title": "FeaturesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Metrics": {
        "properties": {
          "MAE": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae"
          },
          "RMSE": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rmse"
          },
          "R2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R2"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Metrics"
      },
      "ModelInfoResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "model_family": {
            "type": "string",
            "title": "Model Family"
          },
          "package_version": {
            "type": "string",
            "title": "Package Version"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          },
          "feature_set": {
            "type": "string",
            "title": "Feature Set"
          },
          "training_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Date"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_id",
          "model_version",
          "model_family",
          "package_version",
          "data_version",
          "feature_set"
        ],
        "title": "ModelInfoResponse"
      },
      "PredictRequest": {
        "properties": {
          "duration_min": {
            "type": "number",
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Duration Min",
            "description": "Track duration in minutes"
          },
          "explicit": {
            "type": "boolean",
            "title": "Explicit",
            "description": "Whether the track has explicit lyrics"
          },
          "release_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1900.0,
            "title": "Release Year",
            "description": "Release year"
          },
          "release_month": {
            "type": "number",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Release Month",
            "description": "Release month (1–12)"
          },
          "decade": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 1900.0,
            "title": "Decade",
            "description": "Decade bucket"
          },
          "release_precision": {
            "type": "string",
            "pattern": "^(day|month|year)$",
            "title": "Release Precision",
            "description": "Precision of the release date"
          },
          "danceability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Danceability",
            "description": "Danceability score"
          },
          "energy": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Energy",
            "description": "Energy score"
          },
          "key": {
            "type": "integer",
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "Key",
            "description": "Musical key (0–11)"
          },
          "loudness": {
            "type": "number",
            "maximum": 0.0,
            "minimum": -60.0,
            "title": "Loudness",
            "description": "Loudness in dB"
          },
          "mode": {
            "type": "integer",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mode",
            "description": "Mode (0=minor, 1=major)"
          },
          "speechiness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Speechiness",
            "description": "Speechiness score"
          },
          "acousticness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Acousticness",
            "description": "Acousticness score"
          },
          "instrumentalness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Instrumentalness",
            "description": "Instrumentalness score"
          },
          "liveness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Liveness",
            "description": "Liveness score"
          },
          "valence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Valence",
            "description": "Valence score"
          },
          "tempo": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 0.0,
            "title": "Tempo",
            "description": "Tempo in BPM"
          },
          "time_signature": {
            "type": "number",
            "title": "Time Signature",
            "description": "Time signature"
          },
          "CANONICAL_FIELD_NAMES": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Canonical Field Names",
            "default": [
              "acousticness",
              "danceability",
              "decade",
              "duration_min",
              "energy",
              "explicit",
              "instrumentalness",
              "key",
              "liveness",
              "loudness",
              "mode",
              "release_month",
              "release_precision",
              "release_year",
              "speechiness",
              "tempo",
              "time_signature",
              "valence"
            ]
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "duration_min",
          "explicit",
          "release_year",
          "release_month",
          "decade",
          "release_precision",
          "danceability",
          "energy",
          "key",
          "loudness",
          "mode",
          "speechiness",
          "acousticness",
          "instrumentalness",
          "liveness",
          "valence",
          "tempo",
          "time_signature"
        ],
        "title": "PredictRequest",
        "description": "Canonical 18-field input for track popularity prediction.\n\nField names and constraints sourced from:\n- input_schema.json (HITRADAR-PREDICTION-INPUT-V1)\n\nTarget field (target_popularity) and identifier (track_id) are excluded.",
        "example": {
          "acousticness": 0.3,
          "danceability": 0.7,
          "decade": 1990,
          "duration_min": 3.517,
          "energy": 0.8,
          "explicit": false,
          "instrumentalness": 0.05,
          "key": 5,
          "liveness": 0.2,
          "loudness": -5.0,
          "mode": 1,
          "release_month": 1.0,
          "release_precision": "year",
          "release_year": 1992,
          "speechiness": 0.1,
          "tempo": 120.0,
          "time_signature": 4.0,
          "valence": 0.6
        }
      },
...
    }
  }
}